refactor: Split index_ddl.rs into separate CREATE/DROP modules
- Extracted CreateIndexExecutor into dedicated index_ddl/create_index.rs
- Extracted DropIndexExecutor into dedicated index_ddl/drop_index.rs
- Created mod.rs to organize and export the submodules
- Maintained backward compatible IndexExecutor interface
- All 14 index tests pass
- Reduces monolithic file from 526 lines to focused modules (~220 lines each)
- Improved separation of concerns and testability

Closes #1029
