+++ crates/executor/src/index_ddl/create_index.rs
@@ -1,15 +1,15 @@
-//! CREATE INDEX and DROP INDEX statement execution
+//! CREATE INDEX statement execution
 
-use ast::{CreateIndexStmt, DropIndexStmt};
+use ast::CreateIndexStmt;
 use storage::Database;
 
 use crate::errors::ExecutorError;
 use crate::privilege_checker::PrivilegeChecker;
 
 /// Executor for CREATE INDEX statements
-pub struct IndexExecutor;
+pub struct CreateIndexExecutor;
 
-impl IndexExecutor {
+impl CreateIndexExecutor {
     /// Execute a CREATE INDEX statement
     ///
     /// # Arguments
@@ -26,7 +26,7 @@ impl IndexExecutor {
     /// ```
     /// use ast::{CreateIndexStmt, IndexColumn, OrderDirection};
     /// use storage::Database;
-    /// use executor::IndexExecutor;
+    /// use executor::CreateIndexExecutor;
     ///
     /// let mut db = Database::new();
     /// // First create a table
@@ -45,7 +45,7 @@ impl IndexExecutor {
     ///     ],
     /// };
     ///
-    /// let result = IndexExecutor::execute(&stmt, &mut db);
+    /// let result = CreateIndexExecutor::execute(&stmt, &mut db);
     /// // assert!(result.is_ok());
     /// ```
     pub fn execute(
@@ -112,47 +112,14 @@ impl IndexExecutor {
             index_name, qualified_table_name
         ))
     }
-
-    /// Execute a DROP INDEX statement
-    ///
-    /// # Arguments
-    ///
-    /// * `stmt` - The DROP INDEX statement AST node
-    /// * `database` - The database to drop the index from
-    ///
-    /// # Returns
-    ///
-    /// Success message or error
-    pub fn execute_drop(
-        stmt: &DropIndexStmt,
-        database: &mut Database,
-    ) -> Result<String, ExecutorError> {
-        let index_name = &stmt.index_name;
-
-        // Check if index exists
-        if !database.index_exists(index_name) {
-            if stmt.if_exists {
-                // IF EXISTS: silently succeed if index doesn't exist
-                return Ok(format!("Index '{}' does not exist (skipped)", index_name));
-            } else {
-                return Err(ExecutorError::IndexNotFound(index_name.clone()));
-            }
-        }
-
-        // Drop the index
-        database.drop_index(index_name)?;
-
-        Ok(format!("Index '{}' dropped successfully", index_name))
-    }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{CreateTableExecutor, InsertExecutor, SelectExecutor};
-    use ast::{ColumnDef, CreateTableStmt, IndexColumn, OrderDirection, Statement};
-    use types::{DataType, SqlValue};
-    use parser;
+    use crate::CreateTableExecutor;
+    use ast::{ColumnDef, CreateTableStmt, IndexColumn, OrderDirection};
+    use types::DataType;
 
     fn create_test_table(db: &mut Database) {
         let stmt = CreateTableStmt {
@@ -183,7 +150,8 @@ mod tests {
                     comment: None,
                 },
             ],
-            table_constraints: vec![], table_options: vec![],
+            table_constraints: vec![],
+            table_options: vec![],
         };
 
         CreateTableExecutor::execute(&stmt, db).unwrap();
@@ -205,7 +173,7 @@ mod tests {
             }],
         };
 
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_ok());
         assert_eq!(
             result.unwrap(),
@@ -232,7 +200,7 @@ mod tests {
             }],
         };
 
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_ok());
         assert!(db.index_exists("idx_users_email_unique"));
     }
@@ -253,7 +221,7 @@ mod tests {
             ],
         };
 
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_ok());
     }
 
@@ -274,11 +242,11 @@ mod tests {
         };
 
         // First creation succeeds
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_ok());
 
         // Second creation fails
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_err());
         assert!(matches!(result, Err(ExecutorError::IndexAlreadyExists(_))));
     }
@@ -298,7 +266,7 @@ mod tests {
             }],
         };
 
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_err());
         assert!(matches!(result, Err(ExecutorError::TableNotFound(_))));
     }
@@ -319,79 +287,10 @@ mod tests {
             }],
         };
 
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_err());
         assert!(matches!(result, Err(ExecutorError::ColumnNotFound { .. })));
     }
-
-    #[test]
-    fn test_drop_index() {
-        let mut db = Database::new();
-        create_test_table(&mut db);
-
-        // Create index
-        let create_stmt = CreateIndexStmt {
-            index_name: "idx_users_email".to_string(),
-            if_not_exists: false,
-            table_name: "users".to_string(),
-            unique: false,
-            columns: vec![IndexColumn {
-                column_name: "email".to_string(),
-                direction: OrderDirection::Asc,
-            }],
-        };
-        IndexExecutor::execute(&create_stmt, &mut db).unwrap();
-
-        // Drop index
-        let drop_stmt = DropIndexStmt {
-            index_name: "idx_users_email".to_string(),
-            if_exists: false,
-        };
-        let result = IndexExecutor::execute_drop(&drop_stmt, &mut db);
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), "Index 'idx_users_email' dropped successfully");
-
-        // Verify index no longer exists
-        assert!(!db.index_exists("idx_users_email"));
-    }
-
-    #[test]
-    fn test_drop_nonexistent_index() {
-        let mut db = Database::new();
-
-        let drop_stmt = DropIndexStmt {
-            index_name: "nonexistent_index".to_string(),
-            if_exists: false,
-        };
-        let result = IndexExecutor::execute_drop(&drop_stmt, &mut db);
-        assert!(result.is_err());
-        assert!(matches!(result, Err(ExecutorError::IndexNotFound(_))));
-    }
-
-    // TODO: Re-enable this test when SelectExecutor API is stabilized
-    // This test uses private methods that are not accessible from this module
-    // #[test]
-    // fn test_index_based_where_filtering() {
-    //     let mut db = Database::new();
-    //     create_test_table(&mut db);
-    //
-    //     // Create an index on the email column
-    //     let create_index_stmt = CreateIndexStmt {
-    //         index_name: "idx_email".to_string(),
-    //         if_not_exists: false,
-    //         table_name: "users".to_string(),
-    //         unique: false,
-    //         columns: vec![IndexColumn {
-    //             column_name: "email".to_string(),
-    //             direction: OrderDirection::Asc,
-    //         }],
-    //     };
-    //
-    //     IndexExecutor::execute(&create_index_stmt, &mut db).unwrap();
-    //
-    //     // Verify index was created successfully
-    //     assert!(db.index_exists("idx_email"));
-    // }
 
     #[test]
     fn test_create_index_if_not_exists_when_not_exists() {
@@ -409,7 +308,7 @@ mod tests {
             }],
         };
 
-        let result = IndexExecutor::execute(&stmt, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt, &mut db);
         assert!(result.is_ok());
         assert_eq!(
             result.unwrap(),
@@ -434,7 +333,7 @@ mod tests {
                 direction: OrderDirection::Asc,
             }],
         };
-        IndexExecutor::execute(&stmt, &mut db).unwrap();
+        CreateIndexExecutor::execute(&stmt, &mut db).unwrap();
 
         // Second creation with IF NOT EXISTS should succeed
         let stmt_with_if_not_exists = CreateIndexStmt {
@@ -447,80 +346,8 @@ mod tests {
                 direction: OrderDirection::Asc,
             }],
         };
-        let result = IndexExecutor::execute(&stmt_with_if_not_exists, &mut db);
+        let result = CreateIndexExecutor::execute(&stmt_with_if_not_exists, &mut db);
         assert!(result.is_ok());
         assert!(db.index_exists("idx_users_email"));
-    }
-
-    #[test]
-    fn test_drop_index_if_exists_when_exists() {
-        let mut db = Database::new();
-        create_test_table(&mut db);
-
-        // Create index
-        let create_stmt = CreateIndexStmt {
-            index_name: "idx_users_email".to_string(),
-            if_not_exists: false,
-            table_name: "users".to_string(),
-            unique: false,
-            columns: vec![IndexColumn {
-                column_name: "email".to_string(),
-                direction: OrderDirection::Asc,
-            }],
-        };
-        IndexExecutor::execute(&create_stmt, &mut db).unwrap();
-
-        // Drop with IF EXISTS should succeed
-        let drop_stmt = DropIndexStmt {
-            index_name: "idx_users_email".to_string(),
-            if_exists: true,
-        };
-        let result = IndexExecutor::execute_drop(&drop_stmt, &mut db);
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), "Index 'idx_users_email' dropped successfully");
-        assert!(!db.index_exists("idx_users_email"));
-    }
-
-    #[test]
-    fn test_drop_index_if_exists_when_not_exists() {
-        let mut db = Database::new();
-
-        // Drop non-existent index with IF EXISTS should succeed
-        let drop_stmt = DropIndexStmt {
-            index_name: "nonexistent_index".to_string(),
-            if_exists: true,
-        };
-        let result = IndexExecutor::execute_drop(&drop_stmt, &mut db);
-        assert!(result.is_ok());
-        // Silently succeeds when index doesn't exist
-    }
-
-    #[test]
-    fn test_case_insensitive_index_names() {
-        let mut db = Database::new();
-        create_test_table(&mut db);
-
-        // Create index with lowercase name
-        let create_stmt = CreateIndexStmt {
-            index_name: "idx_test".to_string(),
-            if_not_exists: false,
-            table_name: "users".to_string(),
-            unique: false,
-            columns: vec![IndexColumn {
-                column_name: "email".to_string(),
-                direction: OrderDirection::Asc,
-            }],
-        };
-        IndexExecutor::execute(&create_stmt, &mut db).unwrap();
-
-        // Drop with uppercase name should work (normalized to uppercase)
-        let drop_stmt = DropIndexStmt {
-            index_name: "IDX_TEST".to_string(),
-            if_exists: false,
-        };
-        let result = IndexExecutor::execute_drop(&drop_stmt, &mut db);
-        assert!(result.is_ok());
-        assert!(!db.index_exists("idx_test"));
-        assert!(!db.index_exists("IDX_TEST"));
     }
 }
