+++ crates/executor/src/lib.rs
@@ -11,7 +11,7 @@ mod drop_table;
 pub mod errors;
 pub mod evaluator;
 mod grant;
-mod index_ddl;
+pub mod index_ddl;
 mod insert;
 pub mod limits;
 mod optimizer;
@@ -34,7 +34,7 @@ pub use drop_table::DropTableExecutor;
 pub use errors::ExecutorError;
 pub use evaluator::ExpressionEvaluator;
 pub use grant::GrantExecutor;
-pub use index_ddl::IndexExecutor;
+pub use index_ddl::{CreateIndexExecutor, DropIndexExecutor, IndexExecutor};
 pub use insert::InsertExecutor;
 pub use privilege_checker::PrivilegeChecker;
 pub use revoke::RevokeExecutor;
